+++ pkg/ebpftracer/c/headers/common/buffer.h
@@ -476,11 +476,11 @@ statfunc int save_args_to_submit_buf(event_data_t *event, args_t *args)
     return arg_num;
 }
 
-#define events_perf_submit(p, id, ret)        do_perf_submit(&events, p, id, ret, true)
-#define signal_events_perf_submit(p, id, ret) do_perf_submit(&signal_events, p, id, ret, true)
+#define events_ringbuf_submit(p, id, ret)        do_ringbuf_submit(&events, p, id, ret, true, EVENTS_RINGBUF_DISCARD)
+#define signal_events_ringbuf_submit(p, id, ret) do_ringbuf_submit(&signal_events, p, id, ret, true, SIGNAL_EVENTS_RINGBUF_DISCARD)
 
-statfunc int do_perf_submit(void *target, program_data_t *p, u32 id, long ret, bool init_task_ctx)
-{
+statfunc int do_ringbuf_submit(void *target, program_data_t *p, u32 id, long ret, bool init_task_ctx, enum metric m) {
+
     p->event->context.eventid = id;
     p->event->context.retval = ret;
 
@@ -500,7 +500,11 @@ statfunc int do_perf_submit(void *target, program_data_t *p, u32 id, long ret, b
                  :
                  : [size] "r"(size), [max_size] "i"(MAX_EVENT_SIZE));
 
-    return bpf_perf_event_output(p->ctx, target, BPF_F_CURRENT_CPU, p->event, size);
+    if (bpf_ringbuf_output(target, p->event, size, 0) < 0) {
+        metrics_increase(m);
+        return 1;
+    }
+    return 0;
 }
 
 statfunc event_data_t *find_next_free_scratch_buf(void *scratch_map)
