+++ pkg/ebpftracer/metrics_other.go
@@ -1,0 +1,5 @@
+//go:build !linux
+
+package ebpftracer
+
+const ebpfStats = 0
